Code/docker/compose/fog_server/requirements.txt
psycopg2
paho-mqtt
numpy